Add XML comments
