+++ src/Spectre.Console/Internal/Composition/BlockElement.cs
