+++ src/Spectre.Console/Internal/Composition/LineBreakElement.cs
@@ -1,5 +1,4 @@
 using System;
-using Spectre.Console.Internal;
 
 namespace Spectre.Console.Internal
 {
