+++ src/Spectre.Console/Internal/Composition/MarkupElement.cs
