+++ src/Spectre.Console/Internal/Composition/SpaceElement.cs
@@ -1,16 +1,14 @@
-using Spectre.Console.Internal;
-
 namespace Spectre.Console.Internal
 {
-    internal sealed class TabElement : IRenderable
+    internal sealed class SpaceElement : IRenderable
     {
         private readonly RepeatingElement _element;
 
         public int Width => _element.Width;
 
-        public TabElement(int count = 1)
+        public SpaceElement(int count = 1)
         {
-            _element = new RepeatingElement(count * 4, new TextElement(" "));
+            _element = new RepeatingElement(count, new TextElement(" "));
         }
 
         public void Render(IAnsiConsole renderer)
