+++ src/Spectre.Console/Internal/Composition/StyleElement.cs
