+++ src/Spectre.Console/Internal/Composition/TabElement.cs
@@ -1,16 +1,14 @@
-using Spectre.Console.Internal;
-
 namespace Spectre.Console.Internal
 {
-    internal sealed class SpaceElement : IRenderable
+    internal sealed class TabElement : IRenderable
     {
         private readonly RepeatingElement _element;
 
         public int Width => _element.Width;
 
-        public SpaceElement(int count = 1)
+        public TabElement(int count = 1)
         {
-            _element = new RepeatingElement(count, new TextElement(" "));
+            _element = new RepeatingElement(count * 4, new TextElement(" "));
         }
 
         public void Render(IAnsiConsole renderer)
