+++ src/Spectre.Console/Internal/Composition/TextElement.cs
